# orchestrator/docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Fleet orchestrator for the Bittle robot dog. Routes under /api/robots/{robotId} address one robot; most of them relay to the same path on that robot's Python adapter, so their bodies are the adapter's JSON and are typed as plain objects.\n\nErrors use the envelope {\"error\": code, \"message\": text}: 404 robot_not_found, 400 bad_request, 409 behavior_loop_running or behavior_loop_held_elsewhere, and 502 adapter_unavailable when the adapter cannot be reached. Error responses from the adapter itself are forwarded verbatim with their status.\n\nOperations marked deprecated belong to the adapter's old brain and are scheduled for removal.",
    "title" : "Bittle orchestrator API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/robots/{robotId}/behavior" : {
      "get" : {
        "deprecated" : true,
        "description" : "Ollama by default, Claude opt-in.",
        "operationId" : "behavior_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RobotBehavior"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "One behavior decision from the adapter's own engine",
        "tags" : [ "Robot" ]
      }
    },
    "/api/robots/{robotId}/behavior/action/{action}" : {
      "post" : {
        "description" : "Executes immediately when the loop is stopped.",
        "operationId" : "manualAction",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "action",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActionResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Queue a manual action through the loop",
        "tags" : [ "Behavior" ]
      }
    },
    "/api/robots/{robotId}/behavior/event/{type}" : {
      "post" : {
        "description" : "Unknown event names yield 400.",
        "operationId" : "event",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "type",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Snapshot"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Apply a personality event",
        "tags" : [ "Behavior" ]
      }
    },
    "/api/robots/{robotId}/behavior/history" : {
      "get" : {
        "operationId" : "history",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BehaviorDecision"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Recent decisions, newest first",
        "tags" : [ "Behavior" ]
      }
    },
    "/api/robots/{robotId}/behavior/personality" : {
      "get" : {
        "description" : "Six dimensions and posture.",
        "operationId" : "personality_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Snapshot"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Orchestrator personality snapshot",
        "tags" : [ "Behavior" ]
      }
    },
    "/api/robots/{robotId}/behavior/start" : {
      "post" : {
        "description" : "409 behavior_loop_held_elsewhere while another instance holds its lease.",
        "operationId" : "start",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BehaviorStatus"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Start the behavior loop",
        "tags" : [ "Behavior" ]
      }
    },
    "/api/robots/{robotId}/behavior/status" : {
      "get" : {
        "description" : "Loop running, owning instance, personality and last decision.",
        "operationId" : "status_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BehaviorStatus"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Behavior loop status",
        "tags" : [ "Behavior" ]
      }
    },
    "/api/robots/{robotId}/behavior/stop" : {
      "post" : {
        "operationId" : "stop",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BehaviorStatus"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Stop the behavior loop",
        "tags" : [ "Behavior" ]
      }
    },
...
    "/api/robots/{robotId}/behaviors/{name}" : {
      "get" : {
        "description" : "Relayed to the same path on the adapter.",
        "operationId" : "behavior",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "One stored behavior",
        "tags" : [ "Robot" ]
      }
    },
...
    "/api/robots/{robotId}/personality" : {
      "get" : {
        "deprecated" : true,
        "operationId" : "personality",
        "parameters" : [ {
          "in" : "path",
          "name" : "robotId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RobotPersonality"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Adapter-side personality state",
        "tags" : [ "Robot" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ActionResult" : {
        "type" : "object",
        "properties" : {
          "action" : {
            "type" : "string"
          },
          "actualDurationMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "message" : {
            "type" : "string"
          },
          "robotId" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "BehaviorDecision" : {
        "type" : "object",
        "properties" : {
          "actualDurationMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "decidedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reasoning" : {
            "type" : "string"
          },
          "robotId" : {
            "type" : "string"
          },
          "selectedAction" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string",
            "enum" : [ "RULE", "CLAUDE", "MANUAL" ]
          },
          "stateAtDecision" : {
            "$ref" : "#/components/schemas/Snapshot"
          },
          "success" : {
            "type" : "boolean"
          },
          "validActions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "BehaviorStatus" : {
        "type" : "object",
        "properties" : {
          "lastDecision" : {
            "$ref" : "#/components/schemas/BehaviorDecision"
          },
          "owner" : {
            "type" : "string"
          },
          "personality" : {
            "$ref" : "#/components/schemas/Snapshot"
          },
          "robotId" : {
            "type" : "string"
          },
          "running" : {
            "type" : "boolean"
          }
        }
      },
...
      "RobotBehavior" : {
        "type" : "object",
        "properties" : {
          "behavior" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "robotId" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          }
        }
      },
...
      "RobotPersonality" : {
        "type" : "object",
        "properties" : {
          "boredom" : {
            "type" : "number",
            "format" : "double"
          },
          "curiosity" : {
            "type" : "number",
            "format" : "double"
          },
          "energy" : {
            "type" : "number",
            "format" : "double"
          },
          "happiness" : {
            "type" : "number",
            "format" : "double"
          },
          "mood" : {
            "type" : "string"
          },
          "robotId" : {
            "type" : "string"
          }
        }
      },
...
      "Snapshot" : {
        "type" : "object",
        "properties" : {
          "boredom" : {
            "type" : "number",
            "format" : "double"
          },
          "contentment" : {
            "type" : "number",
            "format" : "double"
          },
          "curiosity" : {
            "type" : "number",
            "format" : "double"
          },
          "energy" : {
            "type" : "number",
            "format" : "double"
          },
          "happiness" : {
            "type" : "number",
            "format" : "double"
          },
          "hunger" : {
            "type" : "number",
            "format" : "double"
          },
          "lastAction" : {
            "type" : "string"
          },
          "lastActionAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastInteractionAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "posture" : {
            "type" : "string",
            "enum" : [ "STANDING", "SITTING", "LYING", "SLEEPING" ]
          },
          "robotId" : {
            "type" : "string"
          },
          "totalActionsThisSession" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    }
  }
}